refactor: onSelectedOptionId 타입 변경

+++ backend/src/main/java/reviewme/template/dto/response/SectionResponse.java
@@ -6,7 +6,7 @@ public record SectionResponse(
 
         long sectionId,
         String visible,
-        Integer onSelectedOptionId,
+        Long onSelectedOptionId,
         String header,
         List<QuestionResponse> questions
 ) {
